@@ -30,7 +30,6 @@
     "react-native-modal-datetime-picker": "^17.1.0",
     "react-native-paper": "^5.12.3",
     "react-native-picker-select": "^9.1.3",
-    "react-native-pure-jwt": "^3.0.2",
     "react-native-safe-area-context": "^4.10.1",
     "react-native-screens": "^3.31.1",
     "react-native-vector-icons": "^10.1.0"
@@ -47,7 +46,7 @@
     "eslint": "^8.57.0",
     "eslint-config-universe": "^12.0.1",
     "prettier": "^3.2.5",
-    "typescript": "~5.3.3"
+    "typescript": "^5.5.2"
   },
   "eslintConfig": {
     "extends": "universe/native"
