@@ -5,17 +5,23 @@ import {
   logout,
   saveAccessToken,
   getAccessToken,
-  getAccessTokenPayload,
+  getAccessTokenPayload as fetchAccessTokenPayload,
   isAuthenticated as checkIsAuthenticated,
   hasAnyRoles,
 } from '../service/AuthService';
+
+export interface AccessTokenPayloadDTO {
+  exp: number;
+  user_name: string;
+  authorities: string[]; // Estrutura real das autoridades pode variar
+}
 
 export interface AuthContextType {
   loginRequest: typeof loginRequest;
   logout: typeof logout;
   saveAccessToken: typeof saveAccessToken;
   getAccessToken: typeof getAccessToken;
-  getAccessTokenPayload: typeof getAccessTokenPayload;
+  getAccessTokenPayload: () => Promise<AccessTokenPayloadDTO | null | undefined>; // Permitindo null no retorno
   isAuthenticated: boolean;
   hasAnyRoles: typeof hasAnyRoles;
   loading: boolean;
@@ -39,11 +45,24 @@ export const AuthProvider: React.FC<{ children: ReactNode }> = ({ children }) =>
   useEffect(() => {
     const checkAuth = async () => {
       const authStatus = checkIsAuthenticated();
-      setIsAuthenticated(authStatus);
+      setIsAuthenticated(await authStatus);
       setLoading(false);
     };
     checkAuth();
   }, []);
+
+  const getAccessTokenPayload = async (): Promise<AccessTokenPayloadDTO | null | undefined> => {
+    const token = await getAccessToken();
+    if (token) {
+      try {
+        return await fetchAccessTokenPayload();
+      } catch (error) {
+        console.error('Error decoding token payload:', error);
+        return null; // Retornar null em caso de erro de decodificação
+      }
+    }
+    return undefined; // Retornar undefined se não houver token
+  };
 
   return (
     <AuthContext.Provider
