@@ -1,4 +1,4 @@
-export const BASE_URL = 'http://172.16.90.75:8080';
+export const BASE_URL = 'http://192.168.1.6:8080';
 export const CLIENT_ID = 'myclientid';
 export const CLIENT_SECRET = 'myclientsecret';
 export const TOKEN_KEY = 'com.esibape/token';
