@@ -4,15 +4,18 @@ import { TOKEN_KEY } from '../contexts/system';
 
 export async function save(token: string) {
   try {
-    await AsyncStorage.setItem(TOKEN_KEY, token);
+    await AsyncStorage.setItem('accessToken', token);
   } catch (error) {
     console.error('Error saving token', error);
+    throw error; // Lança o erro para tratamento externo
   }
 }
 
 export async function get(): Promise<string | null> {
   try {
-    return await AsyncStorage.getItem(TOKEN_KEY);
+    const token = await AsyncStorage.getItem('accessToken');
+    console.log('Token recuperadoRepository:', token); // Adicione um log aqui para verificar o token recuperado
+    return token;
   } catch (error) {
     console.error('Error getting token', error);
     return null;
@@ -21,12 +24,12 @@ export async function get(): Promise<string | null> {
 
 export async function remove() {
   try {
-    await AsyncStorage.removeItem(TOKEN_KEY);
+    await AsyncStorage.removeItem('accessToken');
   } catch (error) {
     console.error('Error removing token', error);
+    throw error; // Lança o erro para tratamento externo
   }
 }
-
 export async function getUserData(): Promise<any | null> {
   try {
     const userDataString = await AsyncStorage.getItem('userData');
